requirements.txt
flask
psycopg[binary]
requests
python-dotenv
gunicorn